requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
flask-cors==4.0.0
pymongo==4.6.1
protobuf~=5.29.4
pytest
authlib
requests